+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * [#105](https://github.com/onechiporenko/my-dungeons-book/issues/105) Add Covenant blocks to Roster
 * Track Spiteful Shade melee damage separately
 * [#14](https://github.com/onechiporenko/my-dungeons-book/issues/14) Track party members talents
+* Fix communication between party members
+* Track timeline for idle time 
 
 ### v2.2.0